Projeto/requirements.txt
asgiref==3.8.1
asyncio==3.4.3
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1
filelock==3.16.1
fonttools==4.53.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.0.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.27.1
websockets==12.0
